+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Animal search
 - Verify voluteers
 - Create request on veterinary
+- Admin can see users